requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
# Optional: Uncomment if you want to use Deep Q-Network
# tensorflow>=2.8.0
# or
# torch>=1.10.0